Commit for Nail

+++ MvvMCW/Model/Movie.cs
@@ -7,6 +7,6 @@ namespace MvvMCW.Model
         public string Name { get; set; }
         public string Genre { get; set; }
         public DateTime Year { get; set; }
-        public float Imdb { get; set; }
+        public double Imdb { get; set; }
     }
 }
